Updated person logic

+++ application/controllers/Person.php
@@ -1,46 +1,67 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
+
 class Person extends CI_Controller{
-   public function __construct(){
-       parent::__construct();
-   }
-   public function index()
-   {
-       $data = [
-           'title' => "Person",
-       ];
-      $this->load->view("person", $data);
-   }
-   public function view_person()
-   {
-    $data = [
-        'title' => "View Person",
-    ];
-    $data['persons'] = $this->PersonModel->get_person(); 
-    $this->load->view("view_person", $data);
-   }
+    public function __construct(){
+        parent::__construct();
+    }
+    
+    public function index() {
+        $data = [
+            'title' => "Person",
+            "page" => "person",
+            "persons" => $this->PersonModel->select(),
+        ];
+        $this->load->view("person", $data);
+    }
+
+    public function add(){
+        $data = [
+            'title' => "Person",
+            "page" => "person",
+            "groups" => $this->GroupsModel->select(),
+        ];
+        $this->load->view("person_add", $data);
+    }
+
+    public function view() {
+        $data = [
+            'title' => "View Person",
+            "page" => "person",
+            "groups" => $this->GroupsModel->select(),
+            "person" => $this->PersonModel->select($this->input->get("id")),
+        ];
+        $this->load->view("person_view", $data);
+    }
+
+    public function delete(){
+        $data = [
+            'title' => "Delete Person",
+            "page" => "person",
+            "person" => $this->PersonModel->select($this->input->get("id")),
+        ];
+        $this->load->view("person_delete", $data);
+    }
 
     //this function will add person data.
     public function add_person(){
-        
         if ($this->input->method() != "post"){
             echo utils::response("Invalid request method", "error");
             return false;
         }
         try {
-            echo "<pre/>";
             $this->db->trans_begin();
-            $data = array(
-                'id'=>strip_tags($this->input->post('id')),
-                'type'=>strip_tags($this->input->post('type')),
-                'first_name'=>strip_tags($this->input->post('first_name')),
-                'last_name'=>strip_tags($this->input->post('last_name')),
-                'id_passport'=>strip_tags($this->input->post('id_passport')), 
-                'email' =>strip_tags($this->input->post('email')), 
-                'mobile' =>strip_tags($this->input->post('mobile')), 
-                'date_added' =>date('Y-m-d',strtotime(utils::getDate()))              
-                );
-            $result = $this->PersonModel->add_person($data);
+            // there is no need to define another method to add: check base_model
+            // $result = $this->PersonModel->add_person($data);
+            $this->PersonModel->insert([
+                'type'=> strip_tags($this->input->post('type')),
+                'first_name'=> strip_tags($this->input->post('first_name')),
+                'last_name'=> strip_tags($this->input->post('last_name')),
+                'id_passport'=> strip_tags($this->input->post('id_passport')), 
+                'email' => strip_tags($this->input->post('email')), 
+                'mobile' => strip_tags($this->input->post('mobile')), 
+                'date_added' => strtotime(utils::getDate())
+            ]);
             
             if ($this->db->trans_status() === FALSE) {
                 $this->db->trans_rollback();
@@ -58,40 +79,26 @@ class Person extends CI_Controller{
             return false;
         }
     }
+
     //this function will update the person data
-    public function update_person()
-    {
-        $id = $this->input->get('id');
-        $person = $this->PersonModel->single_person($id);
-        $data = [
-            'title' => "Person",
-            'person' => $person
-        ];
-       
-        //$data['persons'] = $this->Persons->get_person();
-        $this->load->view("update_person", $data); 
-    } 
-    //this function will update the person data
-   public function save_person()
-   {
+    public function save_person(){
         if ($this->input->method() != "post"){
             echo utils::response("Invalid request method", "error");
             return false;
         }
         try {
-            echo "<pre/>";
             $this->db->trans_begin();
-            $id = strip_tags($this->input->post('id'));
-            $data = array(
+
+            $result = $this->PersonModel->update([
+                'id' => strip_tags($this->input->post('id')),
                 'type'=>strip_tags($this->input->post('type')),
                 'first_name'=>strip_tags($this->input->post('first_name')),
                 'last_name'=>strip_tags($this->input->post('last_name')),
                 'id_passport'=>strip_tags($this->input->post('id_passport')), 
                 'email' =>strip_tags($this->input->post('email')), 
                 'mobile' =>strip_tags($this->input->post('mobile')), 
-                'date_added' =>date('Y-m-d',strtotime(utils::getDate()))              
-                );
-            $result = $this->PersonModel->update_person($id, $data);
+                'date_added' =>strtotime(utils::getDate()) 
+            ]);
 
             if ($this->db->trans_status() === FALSE) {
                 $this->db->trans_rollback();
@@ -103,44 +110,36 @@ class Person extends CI_Controller{
                 return true;
             }
         } catch (Exception $e) {
-            print_r($e->getMessage());
             log_message('Error: ', $e->getMessage());
             echo utils::response("Could not register this time. Please try again later!", "error");
             return false;
         }         
-   }
-   public function delete_data()
-   {
-      $id = $this->input->get('id');
-      $person= $this->PersonModel->single_person($id);
-      $result = $this->PersonModel->delete($id);
-      $data = [
-        "title" => "delete Person",
-        "assets" => $this->config->item('assets'),
-        "person" => $person
-      ];   
-     $this->load->view("delete_person", $data);
     }
    
-   public function delete_person($id)
-   {
-        //start the transaction
-        $this->db->trans_begin();
-        //delete person
-        $id=$this->input->get('id');
-	    $this->PersonModel->delete($id);
-        //make transaction complete
-        $this->db->trans_complete();
-        //check if transaction status TRUE or FALSE
-        if ($this->db->trans_status() === FALSE) {
-            //if something went wrong, rollback everything
-            $this->db->trans_rollback();
-        return FALSE;
-        } else {
-            //if everything went right, delete the data from the database
-            $this->db->trans_commit();
-            return TRUE;
+    public function delete_person()
+    {
+        if ($this->input->method() != "post"){
+            echo utils::response("Invalid request method", "error");
+            return false;
         }
+        try {
+            $this->db->trans_begin();
+            $id = strip_tags($this->input->post('id'));
+            $result = $this->PersonModel->delete($id);
+            if ($this->db->trans_status() === FALSE) {
+                $this->db->trans_rollback();
+                echo utils::response("Update failed!", "error");
+                return false;
+            } else {
+                $this->db->trans_commit();
+                echo utils::response(base_url() . "person", "ok");
+                return true;
+            }
+        } catch (Exception $e) {
+            log_message('Error: ', $e->getMessage());
+            echo utils::response("Could not register this time. Please try again later!", "error");
+            return false;
+        }         
     }
 
 }
